+++ src/main/java/io/voucherify/client/model/redemption/RedeemSession.java
@@ -1,0 +1,13 @@
+package io.voucherify.client.model.redemption;
+
+import lombok.*;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Getter
+@Builder
+@ToString
+public class RedeemSession {
+
+    private String key;
+}