+++ src/main/java/io/voucherify/client/model/session/SessionTTLUnit.java
@@ -1,0 +1,9 @@
+package io.voucherify.client.model.session;
+
+public enum SessionTTLUnit {
+    MILLISECONDS,
+    SECONDS,
+    MINUTES,
+    HOURS,
+    DAYS
+}