+++ src/main/java/io/voucherify/client/model/session/SessionType.java
@@ -1,0 +1,5 @@
+package io.voucherify.client.model.session;
+
+public enum SessionType {
+    LOCK
+}